Google Cloud API Project
For this team based project, we successfully created a working data pipeline with the use of soccer, F1, and GDP data which we took from their respective API's.  collaborated with the team and specifically worked with the F1 and GDP API's. You can look further into the ReadMe for further details on the process and purpose of each component.